requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

apache-airflow-providers-dbt-cloud>=3.0.0
astronomer-providers[dbt_cloud]